DOUBLE -> FLOAT

+++ include/feature.h
@@ -2,7 +2,8 @@
 #define __FEATURE_H__
 #include <utility>
 #include <string>
+#include "comm.h"
 int getFeatureOffset(std::string &name);
 int getNumberOfFeature();
-typedef std::pair<int, double> Feature;
+typedef std::pair<int, FLOAT> Feature;
 #endif
